# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API untuk sistem pencatatan absensi sholat siswa",
        "title": "API Sistem Absensi Sholat",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/siswa": {
            "get": {
                "description": "Mengambil daftar siswa dengan dukungan filter, pencarian, sorting, dan pagination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "siswa"
                ],
                "summary": "Ambil semua data siswa dengan filter dan pagination",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cari berdasarkan NIS atau nama siswa",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter berdasarkan kelas (exact match)",
                        "name": "kelas",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter berdasarkan jurusan (exact match)",
                        "name": "jurusan",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter berdasarkan jenis kelamin (Laki-laki/Perempuan)",
                        "name": "jk",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Nomor halaman (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Jumlah item per halaman (default: 20, max: 100)",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Urutkan berdasarkan: nis, nama_siswa, kelas, jurusan (default: nama_siswa)",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Arah pengurutan: asc, desc (default: asc)",
                        "name": "sort_dir",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data siswa berhasil diambil dengan metadata pagination",
                        "schema": {
                            "$ref": "#/definitions/handlers.SiswaListPaginatedResponse"
                        }
                    },
                    "500": {
                        "description": "Kesalahan server internal - Gagal query database",
                        "schema": {
                            "$ref": "#/definitions/handlers.SiswaErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Menambahkan siswa baru ke dalam sistem. NIS harus unik dan belum terdaftar. Semua field (NIS, nama_siswa, jk) wajib diisi",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "siswa"
                ],
                "summary": "Tambah data siswa baru",
                "parameters": [
                    {
                        "description": "Data siswa baru - NIS (unique), nama_siswa, jk (Laki-laki/Perempuan), jurusan (optional), kelas (optional)",
                        "name": "student",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Siswa"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Siswa berhasil ditambahkan - Mengembalikan data siswa yang baru dibuat",
                        "schema": {
                            "$ref": "#/definitions/handlers.SiswaCreateResponse"
                        }
                    },
                    "400": {
                        "description": "Permintaan tidak valid - Format JSON salah, field required kosong, atau duplicate NIS",
                        "schema": {
                            "$ref": "#/definitions/handlers.SiswaErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Kesalahan server internal - Database constraint violation atau connection error",
                        "schema": {
                            "$ref": "#/definitions/handlers.SiswaErrorResponse"
                        }
                    }
                }
            }
        },
        "/siswa/{nis}": {
            "get": {
                "description": "Mengambil detail siswa spesifik berdasarkan nomor induk siswa (NIS). Mengembalikan profil lengkap siswa termasuk nama, jenis kelamin, jurusan, dan kelas",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "siswa"
                ],
                "summary": "Ambil data siswa berdasarkan NIS",
                "parameters": [
                    {
                        "type": "string",
                        "description": "NIS Siswa - nomor induk siswa 4-20 digit",
                        "name": "nis",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data siswa ditemukan dan berhasil diambil",
                        "schema": {
                            "$ref": "#/definitions/handlers.SiswaDetailResponse"
                        }
                    },
                    "400": {
                        "description": "NIS tidak valid atau kosong - parameter path tidak diberikan",
                        "schema": {
                            "$ref": "#/definitions/handlers.SiswaErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Siswa tidak ditemukan - NIS tidak ada dalam database",
                        "schema": {
                            "$ref": "#/definitions/handlers.SiswaErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Kesalahan server internal - Database connection error atau query error",
                        "schema": {
                            "$ref": "#/definitions/handlers.SiswaErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Memperbarui informasi siswa yang sudah terdaftar. NIS tidak dapat diubah (primary key). Hanya nama_siswa, jk, jurusan, dan kelas yang dapat diperbarui",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "siswa"
                ],
                "summary": "Ubah data siswa",
                "parameters": [
                    {
                        "type": "string",
                        "description": "NIS Siswa yang akan diupdate",
                        "name": "nis",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Data siswa yang diperbarui - nama_siswa, jk, jurusan, kelas dapat diubah",
                        "name": "student",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Siswa"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Siswa berhasil diperbarui - Mengembalikan data siswa terbaru",
                        "schema": {
                            "$ref": "#/definitions/handlers.SiswaUpdateResponse"
                        }
                    },
                    "400": {
                        "description": "Permintaan tidak valid - Format JSON salah atau NIS kosong",
                        "schema": {
                            "$ref": "#/definitions/handlers.SiswaErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Siswa tidak ditemukan - NIS tidak ada dalam database",
                        "schema": {
                            "$ref": "#/definitions/handlers.SiswaErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Kesalahan server internal - Database error atau constraint violation",
                        "schema": {
                            "$ref": "#/definitions/handlers.SiswaErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Menghapus siswa dari sistem secara permanen. Operasi ini akan menghapus semua data terkait siswa termasuk akun login dan absensi. Hati-hati karena operasi ini tidak dapat dibatalkan",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "siswa"
                ],
                "summary": "Hapus data siswa",
                "parameters": [
                    {
                        "type": "string",
                        "description": "NIS Siswa yang akan dihapus",
                        "name": "nis",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Siswa berhasil dihapus - Data siswa telah dihapus dari sistem",
                        "schema": {
                            "$ref": "#/definitions/handlers.SiswaDeleteResponse"
                        }
                    },
                    "400": {
                        "description": "Permintaan tidak valid - NIS kosong atau tidak diberikan",
                        "schema": {
                            "$ref": "#/definitions/handlers.SiswaErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Siswa tidak ditemukan - NIS tidak ada dalam database",
                        "schema": {
                            "$ref": "#/definitions/handlers.SiswaErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Kesalahan server internal - Database error atau cascade delete error",
                        "schema": {
                            "$ref": "#/definitions/handlers.SiswaErrorResponse"
                        }
                    }
                }
            }
        },
        "/siswa/{nis}/absensi": {
            "post": {
                "description": "Siswa membuat pencatatan absensi (kehadiran/ketidakhadiran) untuk sesi sholat. Status yang valid adalah: hadir, izin, sakit, alpha",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "absensi"
                ],
                "summary": "Buat absensi sholat baru untuk siswa",
                "parameters": [
                    {
                        "type": "string",
                        "description": "NIS Siswa - nomor induk siswa",
                        "name": "nis",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Data absensi siswa",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateAbsensiRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Absensi berhasil dibuat",
                        "schema": {
                            "$ref": "#/definitions/handlers.AbsensiResponse"
                        }
                    },
                    "400": {
                        "description": "Request tidak valid atau data tidak lengkap",
                        "schema": {
                            "$ref": "#/definitions/handlers.SiswaErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Siswa atau jadwal sholat tidak ditemukan",
                        "schema": {
                            "$ref": "#/definitions/handlers.SiswaErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Kesalahan server internal",
                        "schema": {
                            "$ref": "#/definitions/handlers.SiswaErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.SiswaCreateResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.Siswa"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "handlers.SiswaErrorResponse": {
            "type": "object",
            "properties": {
                "error": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "handlers.SiswaListPaginatedResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Siswa"
                    }
                },
                "filters": {
                    "$ref": "#/definitions/handlers.AppliedFilters"
                },
                "message": {
                    "type": "string"
                },
                "pagination": {
                    "$ref": "#/definitions/handlers.PaginationMeta"
                }
            }
        },
...
        "models.Siswa": {
            "type": "object",
            "properties": {
                "jk": {
                    "type": "string"
                },
                "jurusan": {
                    "type": "string"
                },
                "kelas": {
                    "type": "string"
                },
                "nama_siswa": {
                    "type": "string"
                },
                "nis": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}